Safari-611.1.10/LayoutTests/performance-api/performance-timeline-serializers-expected.txt

Test for performance-timeline serializers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS {"name":"test","entryType":"measure","startTime":"<filtered number>","duration":"<filtered number>"}
PASS Object.keys(JSON.parse(JSON.stringify(entry))).length is 4
PASS {"name":"test","entryType":"mark","startTime":"<filtered number>","duration":"<filtered number>"}
PASS Object.keys(JSON.parse(JSON.stringify(entry))).length is 4
PASS successfullyParsed is true

TEST COMPLETE
